+++ vitest.config.ts
@@ -20,8 +20,5 @@ export default defineConfig({
       '@': resolve(__dirname, 'src/renderer/src'),
       'src': resolve(__dirname, 'src/renderer/src')
     }
-  },
-  define: {
-    __PACKAGE_JSON__: JSON.stringify(require('./package.json'))
   }
 })